refactor(bot): refactor rank command embed
