+++ bathbot/src/embeds/osu/mod.rs
@@ -10,7 +10,6 @@ mod osustats_counts;
 mod player_snipe_stats;
 mod pp_missing;
 mod profile_compare;
-mod rank;
 mod rank_score;
 mod ratio;
 mod sniped;
@@ -25,7 +24,7 @@ pub use self::match_live::*;
 pub use self::{
     attributes::*, bws::*, claim_name::*, country_snipe_stats::*, fix_score::*, match_costs::*,
     medal_stats::*, osustats_counts::*, player_snipe_stats::*, pp_missing::*, profile_compare::*,
-    rank::*, rank_score::*, ratio::*, sniped::*, whatif::*,
+    rank_score::*, ratio::*, sniped::*, whatif::*,
 };
 use crate::manager::OsuMap;
